astrid_tech_api/requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
amqp==1.4.9
anyjson==0.3.3
asgiref==3.3.1; python_version >= '3.5'
billiard==3.3.0.23
bleach==3.2.1
cachetools==4.2.0; python_version ~= '3.5'
celery==3.1.26.post2
certifi==2020.12.5
chardet==4.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
deprecated==1.2.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
django-cached-field==1.5.1
django-celery==3.3.1
django-cors-headers==3.6.0
django-ipware==3.0.2
django-oauth-toolkit==1.3.3
django-rest-framework==0.1.0
django-structlog==2.1.0
django-webmention==2.0.1
django==3.1.4
djangorestframework-recursive==0.1.2
djangorestframework-simplejwt==4.6.0
djangorestframework==3.12.2
google-api-core==1.24.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
google-api-python-client==1.12.8
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.2
google-auth==1.24.0
googleapis-common-protos==1.52.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
gunicorn==20.0.4
httplib2==0.18.1
idna==2.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
kombu==3.0.37
markdown==3.3.3
oauthlib==3.1.0
packaging==20.8; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pillow==8.0.1
protobuf==3.14.0
psycopg2-binary==2.8.6
pyasn1-modules==0.2.8
pyasn1==0.4.8
pygithub==1.54.1
pyjwt==1.7.1
pyparsing==2.4.7; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==0.15.0
python-logstash==0.4.6
pytz==2020.5
requests-oauthlib==1.3.0
requests==2.25.1
rsa==4.6; python_version >= '3.6'
six==1.15.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.1; python_version >= '3.5'
structlog==20.1.0
uritemplate==3.0.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
urllib3==1.26.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
webencodings==0.5.1
wrapt==1.12.1
